-cleanup names, comments and delete double occurences of database stuff
